# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem ujian online (bank soal, jadwal, pengerjaan ujian per section, penilaian).",
        "title": "Ujian Backend API",
        "contact": {
            "name": "Tim Backend Ujian"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/jadwal": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "List jadwal ujian",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all jadwal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Jadwal ujian menghubungkan satu bank soal dengan satu atau lebih kelas peserta, beserta jendela waktu, durasi pengerjaan, dan opsi acak soal/opsi jawaban. Set wajib_token=1 jika peserta harus memasukkan token ujian (lihat GET /ujian-token/current) sebelum bisa mulai mengerjakan; token bersifat opsional per jadwal (default wajib_token=0, tidak perlu token).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Buat jadwal ujian baru",
                "parameters": [
                    {
                        "description": "Data jadwal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.CreateJadwalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create jadwal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
        "/jadwal/aktif/hari-ini": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar jadwal ujian yang berlaku hari ini untuk kelas peserta yang sedang login (berdasarkan JWT), beserta status pengerjaannya (belum mulai/sedang berjalan/selesai).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Jadwal ujian aktif hari ini untuk peserta yang login",
                "responses": {
                    "200": {
                        "description": "Get jadwal aktif hari ini successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalAktifResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/jadwal/bank-soal/{bank_soal_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "List jadwal berdasarkan bank soal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID bank soal (uuid)",
                        "name": "bank_soal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get jadwal by bank soal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/jadwal/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Detail jadwal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID jadwal (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get jadwal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set wajib_token=1 jika peserta harus memasukkan token ujian (lihat GET /ujian-token/current) sebelum bisa mulai mengerjakan; wajib_token=0 berarti jadwal ini tidak butuh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Update jadwal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID jadwal (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data jadwal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.UpdateJadwalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update jadwal successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Hapus jadwal (soft delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID jadwal (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete jadwal successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal menghapus",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/jadwal/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jadwal"
                ],
                "summary": "Restore jadwal yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID jadwal (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore jadwal successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal restore",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_jadwal_dto.CreateJadwalRequest": {
            "type": "object",
            "required": [
                "durasi",
                "id_bank_soal",
                "id_kelas",
                "nama_ujian",
                "tingkat",
                "wkt_mulai",
                "wkt_selesai"
            ],
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "durasi": {
                    "type": "integer",
                    "minimum": 1
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_kelas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nama_ujian": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "wajib_token": {
                    "type": "integer"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JadwalAktifResponse": {
            "type": "object",
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "durasi": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_nilai": {
                    "type": "string"
                },
                "nama_bank_soal": {
                    "type": "string"
                },
                "nama_ujian": {
                    "type": "string"
                },
                "status_pengerjaan": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "wajib_token": {
                    "type": "integer"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JadwalListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JadwalResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JadwalResponse": {
            "type": "object",
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "durasi": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_jurusan": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_jadwal_dto.JurusanItem"
                    }
                },
                "id_kelas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_jadwal_dto.KelasItem"
                    }
                },
                "nama_bank_soal": {
                    "type": "string"
                },
                "nama_ujian": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wajib_token": {
                    "type": "integer"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.JurusanItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "id_jurusan": {
                    "type": "string"
                },
                "nama_jurusan": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.KelasItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "id_kelas": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_jadwal_dto.UpdateJadwalRequest": {
            "type": "object",
            "required": [
                "durasi",
                "id_bank_soal",
                "id_kelas",
                "nama_ujian",
                "tingkat",
                "wkt_mulai",
                "wkt_selesai"
            ],
            "properties": {
                "acak_opsi": {
                    "type": "integer"
                },
                "acak_soal": {
                    "type": "integer"
                },
                "durasi": {
                    "type": "integer",
                    "minimum": 1
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "id_kelas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nama_ujian": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "wajib_token": {
                    "type": "integer"
                },
                "wkt_mulai": {
                    "type": "string"
                },
                "wkt_selesai": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}